Use an object for simulation data and teams
We can store the simulation data in a Javascript object. We create a
blank object and then give it a property called teams, which is at
first an empty array.

The first thing we’ll then do is create the teams, which are themselves
objects. So our simulation data has an array of objects which represent
teams.

Added some helper methods to create the teams (eg. createTeamOne()),
and then some others to help in accessing them (eg. teamOne()).

To loop over the teams in the team array we can use the for … of syntax
of Javascript. `for(var team of teams) { … }` is quite readable.

Instead of dealing with integers to refer to the teams, we can use the
objects that represent the team data.

Clearing the game record is now setting the gameRecord property on each
team object to zero. And when we go to render the game record, we can
look up that gameRecord property for both teams and interpolate the
values into a string. That string is then set as the value of the form
element.

When updating the record, we don’t care which team one. We pass in the
winning team and then we increment that winning team’s gameRecord. And
then we render the game record again, since the value has changed.

This is a small example of how data structures can make the code a bit
easier to use (once familiar with the idea), more explanatory, and more
reusable.

In future commits, it’d be possible to use this practice for other
pieces of data too. Instead of storing the authoritative value in the
form element, we can store it in a javascript object. Updating the
value becomes more straight forward. And updating that form element
only requires reading the value from that object we updated.

Hope this makes sense! Let me know if you have any questions!